+++ site/components/fallback.tsx
@@ -1,5 +1,6 @@
-import 'neko-ui/es/skeleton';
+import { registry, Skeleton } from 'neko-ui';
 
+registry(Skeleton);
 export default function Fallback() {
   return <n-skeleton active={true} title={true} rows={6} />;
 }